LayoutTests/sputnik/Conformance/12_Statement/12.11_The_switch_Statement/S12.11_A3_T4-expected.txt

CONSOLE MESSAGE: line 80: SyntaxError: Unexpected token ':'
S12.11_A3_T4

PASS Expected parsing failure

TEST COMPLETE
